refactor: layouts folder
