+++ packages/user/src/components/ui/avatar.tsx
@@ -2,8 +2,9 @@
 
 import * as React from "react";
 
-import { cn } from "@mlem-user/utils";
 import * as AvatarPrimitive from "@radix-ui/react-avatar";
+
+import { cn } from "@mlem-user/utils";
 
 const Avatar = React.forwardRef<
   React.ElementRef<typeof AvatarPrimitive.Root>,
