+++ packages/user/src/components/ui/badge.tsx
@@ -1,7 +1,8 @@
 import * as React from "react";
 
-import { cn } from "@mlem-user/utils";
 import { cva, type VariantProps } from "class-variance-authority";
+
+import { cn } from "@mlem-user/utils";
 
 const badgeVariants = cva(
   "inline-flex items-center rounded-md border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2",
