+++ packages/user/src/components/ui/dialog.tsx
@@ -1,8 +1,9 @@
 import * as React from "react";
 
-import { cn } from "@mlem-user/utils";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 import { Cross2Icon } from "@radix-ui/react-icons";
+
+import { cn } from "@mlem-user/utils";
 
 const Dialog = DialogPrimitive.Root;
