+++ packages/user/src/components/ui/dropdown-menu.tsx
@@ -2,13 +2,14 @@
 
 import * as React from "react";
 
-import { cn } from "@mlem-user/utils";
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 import {
   CheckIcon,
   ChevronRightIcon,
   DotFilledIcon,
 } from "@radix-ui/react-icons";
+
+import { cn } from "@mlem-user/utils";
 
 const DropdownMenu = DropdownMenuPrimitive.Root;
