+++ packages/user/src/components/ui/sheet.tsx
@@ -2,10 +2,11 @@
 
 import * as React from "react";
 
-import { cn } from "@mlem-user/utils";
 import * as SheetPrimitive from "@radix-ui/react-dialog";
 import { Cross2Icon } from "@radix-ui/react-icons";
 import { cva, type VariantProps } from "class-variance-authority";
+
+import { cn } from "@mlem-user/utils";
 
 const Sheet = SheetPrimitive.Root;
