+++ packages/user/src/modules/@core/mode-toggle/index.tsx
@@ -2,6 +2,9 @@
 
 import * as React from "react";
 
+import { Moon, Sun } from "lucide-react";
+import { useTheme } from "next-themes";
+
 import { Button } from "@mlem-user/components/ui/button";
 import {
   DropdownMenu,
@@ -9,8 +12,6 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@mlem-user/components/ui/dropdown-menu";
-import { Moon, Sun } from "lucide-react";
-import { useTheme } from "next-themes";
 
 export function ModeToggle() {
   const { setTheme } = useTheme();
