+++ packages/user/src/modules/campaign/campaign-detail/WhiteListForm.tsx
@@ -1,13 +1,14 @@
 import { useAccount } from "@casperdash/usewallet";
 import { zodResolver } from "@hookform/resolvers/zod";
+import { useQueryClient } from "@tanstack/react-query";
+import { useForm } from "react-hook-form";
+import * as z from "zod";
+
 import { Button } from "@mlem-user/components/ui/button";
 import { useToast } from "@mlem-user/components/ui/use-toast";
 import { QueryKeys } from "@mlem-user/enums/queryKeys";
 import { useAddUserToWhiteList } from "@mlem-user/services/app/campaign/hooks/useAddUserToWhiteList";
 import { useCheckUserInWhiteList } from "@mlem-user/services/app/campaign/hooks/useCheckUserInWhiteList";
-import { useQueryClient } from "@tanstack/react-query";
-import { useForm } from "react-hook-form";
-import * as z from "zod";
 
 type Props = {
   campaignId: string;
