+++ packages/user/src/services/app/campaign/hooks/useGetCampaigns.ts
@@ -1,5 +1,6 @@
-import { QueryKeys } from "@mlem-user/enums/queryKeys";
 import { useQuery, UseQueryOptions } from "@tanstack/react-query";
+
+import { QueryKeys } from "@mlem-user/enums/queryKeys";
 
 import { getCampaigns } from "..";
 import { GetCampaignsResponse } from "../types";
