+++ packages/user/src/services/app/campaign/hooks/useGetFeaturedCampaigns.ts
@@ -1,6 +1,7 @@
+import { useQuery, UseQueryOptions } from "@tanstack/react-query";
+
 import { QueryKeys } from "@mlem-user/enums/queryKeys";
 import { Campaign } from "@mlem-user/types/campaign";
-import { useQuery, UseQueryOptions } from "@tanstack/react-query";
 
 import { getFeaturedCampaigns } from "..";
